requirements.txt
amqp==2.2.1
awesome-slugify==1.6.5
billiard==3.5.0.3
celery==4.1.0
Django==1.11
django-activeurl==0.1.10
django-appconf==1.0.2
django-autoslug==1.9.3
django-braces==1.11.0
django-classy-tags==0.8.0
django-debug-toolbar==1.8
django-email-extras==0.3.3
django-filter==1.0.4
django-forms-builder==0.13.0
django-groups-manager==0.5.1
django-guardian==1.4.9
django-haystack==2.6.1
django-ipware==1.1.6
django-js-asset==0.1.1
django-kombu==0.9.4
django-material==1.0.0
django-mptt==0.8.7
django-pure-pagination==0.3.0
django-report-builder==3.5.0
django-stdimage==2.4.1
django-stronghold==0.2.9
django-todo==1.6.2
django-viewflow==1.0.2
djangorestframework==3.6.3
et-xmlfile==1.0.1
future==0.15.0
jdcal==1.3
jsonfield==2.0.2
kombu==4.1.0
lxml==3.8.0
mysqlclient==1.3.10
olefile==0.44
openpyxl==2.4.8
Pillow==4.2.1
progressbar2==3.33.2
python-dateutil==2.6.1
python-gnupg==0.4.1
python-utils==2.2.0
pytz==2017.2
regex==2017.7.28
six==1.10.0
sphinx-me==0.3
sqlparse==0.2.3
Unidecode==0.4.21
vine==1.1.4
Whoosh==2.7.4
XlsxWriter==0.9.8
django-filer
git+https://github.com/SmileyChris/easy-thumbnails